+++ be/src/test/java/singleproject/bulletinboard/controller/HomeControllerTest.java
@@ -1,4 +1,4 @@
-package singleproject.bulletinboard;
+package singleproject.bulletinboard.controller;
 
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
@@ -9,7 +9,6 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
-import singleproject.bulletinboard.controller.HomeController;
 
 class HomeControllerTest {
