+++ be/src/test/java/singleproject/bulletinboard/service/MemberServiceTest.java
@@ -1,14 +1,14 @@
 package singleproject.bulletinboard.service;
 
-import static org.junit.jupiter.api.Assertions.*;
 import static org.assertj.core.api.Assertions.*;
 
+import java.time.LocalDateTime;
 import java.util.List;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import singleproject.bulletinboard.controller.RequestJoinMemberInfo;
 import singleproject.bulletinboard.domain.Member;
-import singleproject.bulletinboard.repository.MemberRepository;
 import singleproject.bulletinboard.repository.MemoryMemberRepository;
 
 class MemberServiceTest {
@@ -25,23 +25,24 @@ class MemberServiceTest {
 	@Test
 	void join() {
 		// given
-		Member member1 = Member.builder()
-			.name("철수")
-			.build();
-		memberService.join(member1);
+		RequestJoinMemberInfo joinMemberInfo = new RequestJoinMemberInfo("철수", "123", 25,
+			LocalDateTime.of(1999, 12, 31, 00, 00, 00));
+		memberService.join(joinMemberInfo);
 		// when
 		String member1Name = memberService.findById(1L).orElseThrow().getName();
 		// then
-		assertThat(member1Name).isEqualTo(member1.getName());
+		assertThat(member1Name).isEqualTo(joinMemberInfo.getName());
 	}
 
 	@Test
 	void findMembers() {
 		// given
-		Member member1 = new Member();
-		Member member2 = new Member();
-		memberService.join(member1);
-		memberService.join(member2);
+		RequestJoinMemberInfo joinMemberInfo1 = new RequestJoinMemberInfo("철수", "123", 25,
+			LocalDateTime.of(1999, 12, 31, 00, 00, 00));
+		RequestJoinMemberInfo joinMemberInfo2 = new RequestJoinMemberInfo("영희", "456", 27,
+			LocalDateTime.of(1997, 1, 1, 00, 00, 00));
+		memberService.join(joinMemberInfo1);
+		memberService.join(joinMemberInfo2);
 		// when
 		List<Member> members = memberService.findMembers();
 		// then
@@ -51,20 +52,18 @@ class MemberServiceTest {
 	@Test
 	void findById() {
 		// given
-		Member member1 = Member.builder()
-			.name("철수")
-			.build();
-		Member member2 = Member.builder()
-			.name("영희")
-			.build();
-		memberService.join(member1);
-		memberService.join(member2);
+		RequestJoinMemberInfo joinMemberInfo1 = new RequestJoinMemberInfo("철수", "123", 25,
+			LocalDateTime.of(1999, 12, 31, 00, 00, 00));
+		RequestJoinMemberInfo joinMemberInfo2 = new RequestJoinMemberInfo("영희", "456", 27,
+			LocalDateTime.of(1997, 1, 1, 00, 00, 00));
+		memberService.join(joinMemberInfo1);
+		memberService.join(joinMemberInfo2);
 		// when
 		String member1Name = memberService.findById(1L).orElseThrow().getName();
 		String member2Name = memberService.findById(2L).orElseThrow().getName();
 		// then
-		assertThat(member1.getName()).isEqualTo(member1Name);
-		assertThat(member2.getName()).isEqualTo(member2Name);
+		assertThat(joinMemberInfo1.getName()).isEqualTo(member1Name);
+		assertThat(joinMemberInfo2.getName()).isEqualTo(member2Name);
 	}
 
 	@AfterEach
